bot/requirements.txt
aiogram>=3.4.0,<4
SQLAlchemy>=2.0,<3
python-dotenv>=1.0